+++ src/server/sample.cpp
@@ -524,6 +524,6 @@ awh::server::Sample::Sample(const server::core_t * core, const fmk_t * fmk, cons
 		// Устанавливаем функцию записи данных
 		const_cast <server::core_t *> (this->_core)->on <void (const char *, const size_t, const uint32_t, const uint16_t)> ("write", &sample_t::writeEvent, this, _1, _2, _3, _4);
 		// Добавляем событие аццепта брокера
-		const_cast <server::core_t *> (this->_core)->on <bool (const string &, const string &, const uint32_t, const uint32_t)> ("accept", &sample_t::acceptEvent, this, _1, _2, _3, _4);
+		const_cast <server::core_t *> (this->_core)->on <bool (const string &, const string &, const uint32_t, const uint16_t)> ("accept", &sample_t::acceptEvent, this, _1, _2, _3, _4);
 	}
 }
